+++ Packages/UGF.Module.Descriptions/Runtime/IDescriptionModuleDescription.cs
@@ -1,10 +1,9 @@
 using System.Collections.Generic;
-using UGF.Description.Runtime;
 
 namespace UGF.Module.Descriptions.Runtime
 {
-    public interface IDescriptionModuleDescription : IDescription
+    public interface IDescriptionModuleDescription
     {
-        IReadOnlyList<IDescriptionAssetInfo> AssetInfos { get; }
+        IReadOnlyDictionary<string, string> Assets { get; }
     }
 }